Update README.md
These are protein families somewhat like those found in Pfam.  For some of them, the page may be a source for submission to Pfam.  For others, the emphasis differs in that the intent is to use them as an alignemnt tool for treemaking across the broadest possible range of phages.  This constrasts with the Pfam policy of limiting families to clades sufficiently distinct so as to not give significant similarity scores paralogs or convergently similar sequences.   More information will be provided as the page is populated.